Refactor handlers into separate package
Split the handler functions into a dedicated `handlers` package for improved organization and maintainability. Updated route registrations in `main.go` to reference the new package. Also added `.idea/` to `.gitignore` to exclude IDE-specific files.
