+++ cmd/web/handlers/common.go
@@ -1,0 +1,9 @@
+package handlers
+
+import (
+	"net/http"
+)
+
+func addCommonHeaders(w http.ResponseWriter) {
+	w.Header().Set("X-Server", "Go")
+}
