+++ cmd/web/handlers/homeHandler.go
@@ -1,0 +1,31 @@
+package handlers
+
+import (
+	"html/template"
+	"log"
+	"net/http"
+)
+
+func HomeHandler(w http.ResponseWriter, r *http.Request) {
+	log.Println(r.URL.Path)
+	addCommonHeaders(w)
+
+	files := []string{
+		"ui/html/base.tmpl",
+		"ui/html/partials/nav.tmpl",
+		"ui/html/pages/home.tmpl",
+	}
+
+	ts, err := template.ParseFiles(files...)
+	if err != nil {
+		log.Println(err.Error())
+		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+		return
+	}
+
+	err = ts.ExecuteTemplate(w, "base", nil)
+	if err != nil {
+		log.Println(err.Error())
+		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+	}
+}
