+++ cmd/web/handlers/snippetViewHandler.go
@@ -1,0 +1,20 @@
+package handlers
+
+import (
+	"log"
+	"net/http"
+	"strconv"
+)
+
+func SnippetViewHandler(w http.ResponseWriter, r *http.Request) {
+	log.Println(r.URL.Path)
+	addCommonHeaders(w)
+	// Extract the snippet ID from the URL
+	id, err := strconv.Atoi(r.PathValue("id"))
+	if err != nil || id < 1 {
+		http.NotFound(w, r)
+		return
+	}
+	msg := "Display a specific snippet with ID: " + strconv.Itoa(id)
+	w.Write([]byte(msg))
+}
